+++ app/src/main/java/com/dn0ne/player/app/presentation/components/playback/WavingSeekBar.kt
@@ -233,34 +233,9 @@ fun WavingSeekBar(
             modifier = Modifier.fillMaxWidth(),
             horizontalArrangement = Arrangement.SpaceBetween
         ) {
-            val handlePositionHours =
-                (currentDuration * handleOffsetFraction).toLong() / 1000 / 60 / 60
-            val handlePositionMinutes =
-                (currentDuration * handleOffsetFraction).toLong() / 1000 / 60
-            val handlePositionSeconds =
-                (currentDuration * handleOffsetFraction).toLong() / 1000 % 60
-            val handlePositionText = buildString {
-                if (handlePositionHours > 0) {
-                    append("$handlePositionHours".padStart(2, '0'))
-                    append(":")
-                }
-                append("$handlePositionMinutes".padStart(2, '0'))
-                append(":")
-                append("$handlePositionSeconds".padStart(2, '0'))
-            }
-
-            val durationHours = currentDuration / 1000 / 60 / 60
-            val durationMinutes = currentDuration / 1000 / 60
-            val durationSeconds = currentDuration / 1000 % 60
-            val durationText = buildString {
-                if (durationHours > 0) {
-                    append("$durationHours".padStart(2, '0'))
-                    append(":")
-                }
-                append("$durationMinutes".padStart(2, '0'))
-                append(":")
-                append("$durationSeconds".padStart(2, '0'))
-            }
+            val handlePositionText =
+                formatPlaybackTime((currentDuration * handleOffsetFraction).toLong())
+            val durationText = formatPlaybackTime(currentDuration)
 
             Text(
                 text = handlePositionText,